git push

+++ backup_summary.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
-
 # Variáveis para armazenar os modos e opções
-checkMode=false;
-tfile="";
-regex="";
+checkMode=false
+tfile=""
+regex=""
+fileList=()
 
 # Processamento das opções passadas na linha de comando
 while getopts ":cb:r:" opt; do
@@ -18,6 +18,16 @@ case ${opt} in
     b)
         # Define o ficheiro de exclusão
         tfile="$OPTARG"
+
+        if [ -n "$tfile" ] && [ -f "$tfile" ]; then
+
+            while read -r LINE; do
+
+                fileList+=("$LINE")
+
+            done < "$tfile"
+
+        fi
 
     ;;
 
@@ -35,25 +45,15 @@ case ${opt} in
 
     ;;
 
-esac
+    esac
 done
 
 # Função para verificar se um ficheiro corresponde à expressão regular
 function regexM(){
     if [ -n "$regex" ] && [[ ! "$1" =~ $regex ]]; then
 
-        return 1;
-
-    fi
-    return 0;
-}
-
-# Função para verificar se um ficheiro deve ser excluído
-function fileM(){
-  if [ -n "$tfile" ] && grep -qxF "$(basename "$1")" "$tfile"; then
-        
         return 1
-
+    
     fi
     return 0
 }
@@ -62,7 +62,7 @@ function fileM(){
 function checkModeM(){
     if [ $checkMode == false ]; then
 
-        "$@";
+        "$@"
 
     fi
 }
@@ -70,8 +70,28 @@ function checkModeM(){
 # Remove as opções processadas da lista de argumentos.
 shift $((OPTIND - 1)) 
 
-pathtoDir="$1";
-backupDir="$2";
+pathtoDir="$1"
+backupDir="$2"
+
+# Função para verificar se um ficheiro deve ser excluído
+function fileM() {
+    for item in "${fileList[@]}"; do
+        if [[ "$(basename "$1")" =~ "$item" ]]; then
+            
+            return 1  # O ficheiro está na lista de exclusão
+
+        fi
+    done
+    return 0
+}
+
+if [ ! -d "$pathtoDir" ]; then
+
+    echo "Error: O diretório de trabalho '$pathtoDir' não existe."
+
+    exit 1
+
+fi
 
 # Função principal para realizar o backup
 function accsBackup(){
@@ -81,84 +101,81 @@ function accsBackup(){
 
         echo "Creating Backup Directory"
 
-        checkModeM mkdir -p "$backupDir";
+        checkModeM mkdir -p "$backupDir"
 
-        echo "mkdir -p $backupDir";
+        echo "mkdir -p $backupDir"
     
     else
     
-        echo "Backup Directory Already Exists";
+        echo "Backup Directory Already Exists"
     
     fi
 
     # Verifica se o diretório de backup está vazio
     if find $backupDir -empty -type d; then
 
-            checkModeM cp -a $pathtoDir/. $backupDir;
+            checkModeM cp -a $pathtoDir/. $backupDir
 
-            echo "cp -a $pathtoDir/. $backupDir";
+            echo "cp -a $pathtoDir/. $backupDir"
 
     else
 
-        echo "Files that are in the Directory we want to backup";
+        echo "Files that are in the Directory we want to backup"
 
-        checkModeM ls -l $pathtoDir;
+        checkModeM ls -l $pathtoDir
 
-        echo "Files in the Backup Directory";
+        echo "Files in the Backup Directory"
 
-        checkModeM ls -l $backupDir;
+        checkModeM ls -l $backupDir
 
-        RecursiveDir "$pathtoDir" "$backupDir";
+        RecursiveDir "$pathtoDir" "$backupDir"
 
     fi
 }
 
 # Função recursiva para copiar arquivos e diretórios
 function RecursiveDir(){
-
     for file in $pathtoDir/*; do
 
-        backup_file="$backupDir/$(basename "$file")";
+        echo "$file"
 
-        # Verifica se o arquivo não deve ser excluído e se corresponde à expressão regular
-        if [ fileM "$file" = 0 ]  && [ regexM "$file" ]; then
+        backup_file="$backupDir/$(basename "$file")"
         
-            if [ -f "$backup_file" ]; then
+        if [ -f "$backup_file" ] &&  fileM "$backup_file" ; then
 
-                date_file=$(ls -l "$file" | awk '{print $6}');
+            date_file=$(ls -l "$file" | awk '{print $6}')
 
-                backup_date=$(ls -l "$backup_file" | awk '{print $6}');
+            backup_date=$(ls -l "$backup_file" | awk '{print $6}')
 
-                if [ "$date_file" == "$backup_date" ]; then
+            if [ "$date_file" == "$backup_date" ]; then
 
-                    echo "File $(basename "$file") is up-to-date."
-
-                else
-
-                    echo "File $(basename "$file") has a different modification date.";
-
-                    checkModeM cp -a "$file" $backupDir;
-
-                    echo "cp -a "$file" $backupDir" ;
-
-                fi
-
-            elif [ -d "$file" ]; then
-
-                    checkModeM cp -a "$file" $backupDir;
-
-                    echo "cp -a "$file" $backupDir" ;
-
-                    RecursiveDir "$file" "$backup_file";
+                echo "File $(basename "$file") is up-to-date."
 
             else
 
-                echo "File $(basename "$file") is missing. Let's add it to the backup.";
+                echo "File $(basename "$file") has a different modification date."
 
-                checkModeM cp -a "$file" $backupDir;
+                checkModeM cp -a "$file" $backupDir
 
-                echo "cp -a "$file" $backupDir" ;
+                echo "cp -a "$file" $backupDir" 
+
             fi
+
+        elif [ -d "$file" ]; then
+
+                checkModeM cp -a "$file" $backupDir
+
+                echo "cp -a "$file" $backupDir" 
+
+                RecursiveDir "$file" "$backup_file"
+
+        else
+
+            echo "File $(basename "$file") is missing. Let's add it to the backup."
+
+            checkModeM cp -a "$file" $backupDir
+
+            echo "cp -a "$file" $backupDir" 
         fi
 
     done
